requirements.txt
Babel==2.6.0
Click==7.0
Flask==1.1.1
Flask-Babel==0.11.2
Frozen-Flask==0.15
itsdangerous==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.0
pdfkit==0.6.1
pytz==2018.9
PyYAML==5.3
Werkzeug==0.16.0
